+++ src/constants/entityTypeTranslationKeys.js
@@ -1,9 +1,0 @@
-import entityTypes from './entityTypes';
-
-const entityTypeTranslationKeys = {
-  [entityTypes.AGREEMENT]: 'ui-agreements.agreement',
-  [entityTypes.AGREEMENT_LINE]: 'ui-agreements.agreementLine',
-  [entityTypes.ERESOURCE]: 'ui-agreements.eresource',
-};
-
-export default entityTypeTranslationKeys;