+++ src/routes/UrlCustomizerCreateRoute/UrlCustomizerCreateRoute.js
@@ -8,7 +8,8 @@ import { useCallout, useOkapiKy } from '@folio/stripes/core';
 
 import View from '../../components/views/UrlCustomizerForm';
 import { urls } from '../../components/utilities';
-import { STRING_TEMPLATES_ENDPOINT } from '../../constants/endpoints';
+
+import { STRING_TEMPLATES_ENDPOINT } from '../../constants';
 
 const UrlCustomizerCreateRoute = ({
   handlers,