+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * STRIPES-870 BREAKING upgrade react to v18
   * ERM-2991 Upgrade ui-agreements React to v18
 * ERM-2983 Added documents filter to Agreements search and sort
+* ERM-3012 Agreement relationship type translations don't show in Agreements UI
+* ERM-3027 *BREAKING* bump `react-intl` to `v6.4.4`
+
 ## 9.0.0 2023-02-22
 * ERM-2613 Remove unneeded `react-redux`. Upgrade `stripes-acq-components` to `v4`
 * ERM-2596 Increment ui-agreements to Stripes v8